ref : EditCurrentPriceModal
